remove old plot
